put light in its own file and add a transform class
