+++ renderlib/CMakeLists.txt
@@ -31,12 +31,16 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/IFileReader.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/ImageXYZC.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/ImageXYZC.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Light.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Light.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Logging.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Logging.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Manipulator.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Manipulator.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/MathUtil.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/MathUtil.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Object3d.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Object3d.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RenderSettings.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RenderSettings.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RotateTool.cpp"
