+++ renderlib/Object3d.cpp
@@ -1,0 +1,1 @@
+#include "Object3d.h"
